Implemented console arg parsing

+++ ChuLife/ChuLifeBackend/Consts.hpp
@@ -4,8 +4,8 @@ static constexpr char DEAD_CELL = '0';
 static constexpr char LIFE_CELL = '1';
 static constexpr char ERR_CELL  = '2';
 
-static constexpr unsigned DEFAULT_LIFE_LIMIT = 5u;
-static constexpr unsigned DEFAULT_DEAD_LIMIT = 3u;
+static constexpr unsigned long long DEFAULT_LIFE_LIMIT = 5ull;
+static constexpr unsigned long long DEFAULT_DEAD_LIMIT = 3ull;
 
 static constexpr int DEFAULT_WINDOW_HEIGHT = 720;
 static constexpr int DEFAULT_WINDOW_WIDTH  = 1280;
